chore(VExpansionPanels): rename component folder
