+++ packages/vuetify/src/components/VExpansionPanels/VExpansionPanel.sass
