+++ packages/vuetify/src/components/VExpansionPanels/VExpansionPanel.ts
