+++ packages/vuetify/src/components/VExpansionPanels/VExpansionPanelContent.ts
