+++ packages/vuetify/src/components/VExpansionPanels/VExpansionPanelHeader.ts
