+++ packages/vuetify/src/components/VExpansionPanels/VExpansionPanels.ts
