+++ packages/vuetify/src/components/VExpansionPanels/__tests__/VExpansionPanel.spec.ts
