+++ packages/vuetify/src/components/VExpansionPanels/__tests__/VExpansionPanelHeader.spec.ts
