+++ packages/vuetify/src/components/VExpansionPanels/__tests__/VExpansionPanels.spec.ts
