+++ packages/vuetify/src/components/VExpansionPanels/_variables.scss
