+++ packages/vuetify/src/components/VExpansionPanels/index.ts
