+++ packages/vuetify/src/components/index.ts
@@ -26,7 +26,7 @@ export * from './VDataTable'
 export * from './VDatePicker'
 export * from './VDialog'
 export * from './VDivider'
-export * from './VExpansionPanel'
+export * from './VExpansionPanels'
 export * from './VFileInput'
 export * from './VFooter'
 export * from './VForm'
